+++ ballontranslator/ui/scene_textlayout.py
